[FEAT] #15 예시로 usecase 적용 코드 추가
