+++ Pinit/Pinit/Resources/DIContainer.swift
@@ -1,0 +1,24 @@
+//
+//  DIContainer.swift
+//  Pinit
+//
+//  Created by 안정흠 on 3/18/25.
+//
+
+import Foundation
+import CoreData
+final class DIContainer: NSObject {
+    static let usecase: UseCase = {
+        let container = NSPersistentContainer(name: "Pinit")
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        // context.perform을 제대로 사용하기위해 백그라운드컨텍스트로 가져옴
+        // 비동기 작업에 사용됌
+        let context = container.newBackgroundContext()
+        return UseCaseImpl(dbRepository: DBRepositoryImpl(context: context),
+                    imageStore: ImageStoreRepositoryImpl())
+    }()
+}
