+++ Pinit/Pinit/Views/MainTabBarController.swift
@@ -9,19 +9,6 @@ import UIKit
 import CoreData
 
 final class MainTabBarController: UITabBarController {
-    private lazy var usecase: UseCase = {
-        let container = NSPersistentContainer(name: "Pinit")
-        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        })
-        // context.perform을 제대로 사용하기위해 백그라운드컨텍스트로 가져옴
-        // 비동기 작업에 사용됌
-        let context = container.newBackgroundContext()
-        return UseCaseImpl(dbRepository: DBRepositoryImpl(context: context),
-                    imageStore: ImageStoreRepositoryImpl())
-    }()
     override func viewDidLoad() {
         super.viewDidLoad()
         setupViewControllers()
@@ -58,7 +45,7 @@ final class MainTabBarController: UITabBarController {
     }
     
     private func setupViewControllers() {
-        let home = UINavigationController(rootViewController: HomeViewController())
+        let home = UINavigationController(rootViewController: HomeViewController(usecase: DIContainer.usecase))
         home.tabBarItem = UITabBarItem(title: "Home",
                                        image: UIImage(systemName: "house"),
                                        selectedImage: UIImage(systemName: "house.fill"))
